Добавил получение списка займов от API и проверку начального экрана. В зависимости от наличия токена начальный экран будет авторизация или список займов

+++ app/src/main/java/com/homework/homeworkkuritsyn/data/network/models/Loan.kt
@@ -4,7 +4,7 @@ import java.math.BigDecimal
 import java.math.BigInteger
 
 data class Loan(
-    val amount: BigInteger,
+    val amount: BigDecimal,
     val date: String,
     val firstName: String,
     val id: Int,
